Changed Structure of the repo
