+++ .gitignore
@@ -1,2 +1,3 @@
 
 export_params/params.json
+.spyproject/*
