+++ src/train_models.py
@@ -1,4 +1,4 @@
-'''
+"""
 Copyright 2020 Google LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,10 +12,10 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-'''
+"""
 
 import tensorflow as tf
-import json 
+import json
 INPUT = ['patch'+s for s in 'BGRNP']
 OUTPUT = ['troposhpheric_NO2_column_number_density']
 FEATURES = INPUT+OUTPUT
