+++ src/training/cnn.py
